sensehat_cloudrun/requirements.txt
flask==3.1.0
google-cloud-bigquery==3.25.0
gunicorn==23.0.0
